requirements.txt
botbuilder-integration-aiohttp>=4.14.0
python-dotenv
openai
gunicorn
aiohttp
botbuilder
